Extent-class {raster}	R Documentation

Class "Extent"

Description

Objects of class Extent are used to define the spatial extent (extremes) of objects of the BasicRaster and Raster* classes.

Objects from the Class

You can use the extent function to create Extent objects, or to extract them from Raster* and Spatial* objects.

Slots

xmin:

minimum x coordinate

xmax:

maximum x coordinate

ymin:

minumum y coordinate

ymax:

maximum y coordinate

Methods

show

display values of a Extent object

See Also

extent, setExtent

Examples


ext <- extent(-180,180,-90,90)
ext

[Package raster version 3.5-2 Index]