Summary-methods {raster}	R Documentation

Summary methods

Description

The following summary methods are available for Raster* objects:

mean, median, max, min, range, prod, sum, any, all

All methods take na.rm as an additional logical argument. Default is na.rm=FALSE. If TRUE, NA values are removed from calculations. These methods compute a summary statistic based on cell values of RasterLayers and the result of these methods is always a single RasterLayer (except for range, which returns a RasterBrick with two layers). See calc for functions not included here (e.g. median) or any other custom functions.

You can mix RasterLayer, RasterStack and RasterBrick objects with single numeric or logical values. However, because generic functions are used, the method applied is chosen based on the first argument: 'x'. This means that if r is a RasterLayer object, mean(r, 5) will work, but mean(5, r) will not work.

To summarize all cells within a single RasterLayer, see cellStats and maxValue and minValue

Value

a RasterLayer

See Also

calc

Examples


r1 <- raster(nrow=10, ncol=10)
r1 <- setValues(r1, runif(ncell(r1)))
r2 <- setValues(r1, runif(ncell(r1)))
r3 <- setValues(r1, runif(ncell(r1)))
r <- max(r1, r2, r3)
r <- range(r1, r2, r3, 1.2)

s <- stack(r1, r2, r3)
r <- mean(s, 2)

[Package raster version 3.5-2 Index]