Aggregate raster cells or SpatialPolygons/Lines

Raster* objects:

Aggregate a Raster* object to create a new RasterLayer or RasterBrick with a lower resolution (larger cells). Aggregation groups rectangular areas to create larger cells. The value for the resulting cells is computed with a user-specified function.

SpatialPolygon*:

Aggregate a SpatialPolygon* object, optionally by combining polygons that have the same attributes for one or more variables. If the polygons touch or overlap, internal boundaries are optionally "dissolved".

# S4 method for Raster
aggregate(x, fact, fun=mean, expand=TRUE, na.rm=TRUE, filename='', ...)

# S4 method for SpatialPolygons
aggregate(x, by, sums, dissolve=TRUE, vars=NULL, ...)

Arguments

x

Raster* object or SpatialPolygons* object

fact

postive integer. Aggregation factor expressed as number of cells in each direction (horizontally and vertically). Or two integers (horizontal and vertical aggregation factor) or three integers (when also aggregating over layers). See Details

fun

function used to aggregate values

expand

logical. If TRUE the output Raster* object will be larger than the input Raster* object if a division of the number of columns or rows with factor is not an integer

na.rm

logical. If TRUE, NA cells are removed from calculations

filename

character. Output filename (optional)

...

if x is a Raster* object, additional arguments as for writeRaster

by

character or integer. The variables (column names or numbers) that should be used to aggregate (dissolve) the SpatialPolygons by only maintaining unique combinations of these variables. The default setting is to use no variables and aggregate all polygons. You can also supply a vector with a length of length(x)

sums

list with function(s) and variable(s) to summarize. This should be a list of lists in which each element of the main lists has two items. The first item is function (e.g. mean), the second element is a vector of column names (or indices) that need to summarize with that function. Be careful with character and factor variables (you can use, e.g. 'first' function(x)x[1] or 'last' function(x)x[length(x)] or modal for these variables

vars

deprecated. Same as by

dissolve

logical. If TRUE borders between touching or overlapping polygons are removed

Details

Aggregation of a x will result in a Raster* object with fewer cells. The number of cells is the number of cells of x divided by fact*fact (when fact is a single number) or prod(fact) (when fact consists of 2 or 3 numbers). If necessary this number is adjusted according to the value of expand. For example, fact=2 will result in a new Raster* object with 2*2=4 times fewer cells. If two numbers are supplied, e.g., fact=c(2,3), the first will be used for aggregating in the horizontal direction, and the second for aggregating in the vertical direction, and the returned object will have 2*3=6 times fewer cells. Likewise, fact=c(2,3,4) aggregates cells in groups of 2 (rows) by 3 (columns) and 4 (layers).

Aggregation starts at the upper-left end of a raster (you can use flip if you want to start elsewhere). If a division of the number of columns or rows with factor does not return an integer, the extent of the resulting Raster object will either be somewhat smaller or somewhat larger than the original RasterLayer. For example, if an input RasterLayer has 100 columns, and fact=12, the output Raster object will have either 8 columns (expand=FALSE) (using 8 x 12 = 96 of the original columns) or 9 columns (expand=TRUE). In both cases, the maximum x coordinate of the output RasterLayer would, of course, also be adjusted.

The function fun should take multiple numbers, and return a single number. For example mean, modal, min or max. It should also accept a na.rm argument (or ignore it as one of the 'dots' arguments).

Value

RasterLayer or RasterBrick, or a SpatialPolygons* object

See also

Author

Robert J. Hijmans and Jacob van Etten

Examples

r <- raster()
# a new aggregated raster, no values
ra <- aggregate(r, fact=10)
r <- setValues(r, runif(ncell(r)))

# a new aggregated raster, max of the values
ra <- aggregate(r, fact=10, fun=max)

# multiple layers
s <- stack(r, r*2)
x <- aggregate(s,2)

#SpatialPolygons
p <- shapefile(system.file("external/lux.shp", package="raster"))
p
#> class       : SpatialPolygonsDataFrame 
#> features    : 12 
#> extent      : 5.74414, 6.528252, 49.44781, 50.18162  (xmin, xmax, ymin, ymax)
#> crs         : +proj=longlat +datum=WGS84 +no_defs 
#> variables   : 5
#> names       : ID_1,     NAME_1, ID_2,   NAME_2, AREA 
#> min values  :    1,   Diekirch,    1, Capellen,   76 
#> max values  :    3, Luxembourg,   12,    Wiltz,  312 
pa0 <- aggregate(p)
pa0
#> class       : SpatialPolygons 
#> features    : 1 
#> extent      : 5.74414, 6.528252, 49.44781, 50.18162  (xmin, xmax, ymin, ymax)
#> crs         : +proj=longlat +datum=WGS84 +no_defs 
pa1 <- aggregate(p, by='NAME_1', sums=list(list(mean, 'ID_2')))
pa1
#> class       : SpatialPolygonsDataFrame 
#> features    : 3 
#> extent      : 5.74414, 6.528252, 49.44781, 50.18162  (xmin, xmax, ymin, ymax)
#> crs         : +proj=longlat +datum=WGS84 +no_defs 
#> variables   : 2
#> names       :     NAME_1, ID_2 
#> min values  :   Diekirch,    3 
#> max values  : Luxembourg,  9.5