alignExtent {raster}	R Documentation

Align an extent (object of class Extent)

Description

Align an Extent object with the (boundaries of the) cells of a Raster* object

Usage


alignExtent(extent, object, snap='near')

Arguments

extent
Extent object

object
Raster* object

snap
Character. One of 'near', 'in', or 'out', to determine in which direction the extent should be aligned. To the nearest border, inwards or outwards

Details

Aligning an Extent object to another object assures that it gets the same origin and resolution. This should only be used to adjust objects because of imprecision in the data. alignExtent should not be used to force data to match that really does not match (use e.g. resample or (dis)aggregate for this).

Value

Extent object

See Also

extent, drawExtent, Extent-class

Examples


r <- raster()
e <- extent(-10.1, 9.9, -20.1, 19.9)
ea <- alignExtent(e, r)
e
extent(r)
ea


[Package raster version 3.5-2 Index]