atan2 {raster}	R Documentation

Two argument arc-tangent

Description

For Raster* objects x and y, atan2(y, x) returns the angle in radians for the tangent y/x, handling the case when x is zero. See Trig

See Math-methods for other trigonometric and mathematical functions that can be used with Raster* objects.

Usage


atan2(y, x) 

Arguments

y
Raster* object

x
Raster* object

See Also

Math-methods

Examples


r1 <- r2 <- raster(nrow=10, ncol=10)
values(r1) <- (runif(ncell(r1))-0.5) * 10
values(r2) <- (runif(ncell(r1))-0.5) * 10
atan2(r1, r2)

[Package raster version 3.5-2 Index]