barplot {raster}	R Documentation

Bar plot of a RasterLayer

Description

Create a barplot of the values of a RasterLayer. For large datasets a regular sample with a size of approximately maxpixels is used.

Usage


## S4 method for signature 'RasterLayer'
barplot(height, maxpixels=1000000, digits=0, breaks=NULL, col=rainbow, ...)

Arguments

height
RasterLayer

maxpixels
integer. To regularly subsample very large objects

digits
integer used to determine how to round the values before tabulating. Set to NULL or to a large number if you do not want any rounding

breaks
breaks used to group the data as in cut

col
a color generating function such as rainbow, or a vector of colors

...
additional arguments for plotting as in barplot

Value

A numeric vector (or matrix, when beside = TRUE) of the coordinates of the bar midpoints, useful for adding to the graph. See barplot

See Also

hist, boxplot

Examples

 
f <- system.file("external/test.grd", package="raster")
r <- raster(f)
barplot(r, digits=-2, las=2, ylab='Frequency')

op <- par(no.readonly = TRUE)
par(mai = c(1, 2, .5, .5))
barplot(r, breaks=10, col=c('red', 'blue'), horiz=TRUE, digits=NULL, las=1)
par(op)

[Package raster version 3.5-2 Index]