blockSize {raster}	R Documentation

Block size for writing files

Description

This function can be used to suggest chunk sizes (always a number of entire rows), and corresponding row numbers, to be used when processing Raster* objects in chunks. Normally used together with writeValues.

Usage


blockSize(x, chunksize, n=nlayers(x), minblocks=4, minrows=1)

Arguments

x
Raster* object

chunksize
Integer, normally missing. Can be used to set the block size; unit is number of cells. Block size is then computed in units of number of rows (always >= 1)

n
Integer. number of layers to consider. The function divides chunksize by n to determine blocksize

minblocks
Integer. Minimum number of blocks

minrows
Integer. Minimum number of rows in each block

Value

A list with three elements:

rows, the suggested row numbers at which to start the blocks for reading and writing,

nrows, the number of rows in each block, and,

n, the total number of blocks

See Also

writeValues

Examples


r <- raster(system.file("external/test.grd", package="raster"))
blockSize(r)

[Package raster version 3.5-2 Index]