cellsFromExtent {raster}	R Documentation

Cells from extent, and vice versa

Description

cellsFromExtent returns the cell numbers for a Raster* object that are within a specfied extent (rectangular area), supply an object of class Extent, or another Raster* object.

extentFromCells returns an Extent object from a Raster* object and cell numbers. All cells are within the returned Extent.

Usage


cellsFromExtent(object, extent, expand=FALSE)
extentFromCells(object, cells)

Arguments

object
A Raster* object

extent
An object of class Extent (which you can create with newExtent(), or another Raster* object )

expand
Logical. If TRUE, NA is returned for (virtual) cells implied by bndbox, that are outside the RasterLayer (object). If FALSE, only cell numbers for the area where object and bndbox overlap are returned (see intersect)

cells
numeric. A vector of cell numbers

Value

a vector of cell numbers

See Also

extent, cellFromXY

Examples


r <- raster()
bb <- extent(-5, 5, -5, 5)
cells <- cellsFromExtent(r, bb)
r <- crop(r, bb)
values(r) <- cells

e <- extentFromCells(r, 50:55)

[Package raster version 3.5-2 Index]