contour {raster}	R Documentation

Contour plot

Description

Contour plot of a RasterLayer.

Usage


## S4 method for signature 'RasterLayer'
contour(x, maxpixels=100000, ...)

Arguments

x
Raster* object

maxpixels
maximum number of pixels used to create the contours

...
any argument that can be passed to contour (graphics package)

See Also

persp, filledContour, rasterToContour

Examples


r <- raster(system.file("external/test.grd", package="raster"))
plot(r)
contour(r, add=TRUE)

[Package raster version 3.5-2 Index]