crosstab {raster}	R Documentation

Cross-tabulate

Description

Cross-tabulate two RasterLayer objects, or mulitiple layers in a RasterStack or RasterBrick to create a contingency table.

Usage


## S4 method for signature 'Raster,Raster'
crosstab(x, y, digits=0, long=FALSE, useNA=FALSE, progress='', ...)

## S4 method for signature 'RasterStackBrick,missing'
crosstab(x, digits=0, long=FALSE, useNA=FALSE, progress='', ...)

Arguments

x
Raster* object

y
Raster* object if x is a RasterLayer; Can be missing if x is a RasterStack or RasterBrick

digits
integer. The number of digits for rounding the values before cross-tabulation

long
logical. If TRUE the results are returned in 'long' format data.frame instead of a table

useNA
logical, indicting if the table should includes counts of NA values

progress
character. "text", "window", or "" (the default, no progress bar), only for large files that cannot be processed in one step

...
additional arguments. none implemented

Value

A table or data.frame

See Also

freq, zonal

Examples


r <- raster(nc=5, nr=5)
values(r) <- runif(ncell(r)) * 2
s <- setValues(r, runif(ncell(r)) * 3)
crosstab(r,s)

rs <- r/s
r[1:5] <- NA
s[20:25] <- NA
x <- stack(r, s, rs)
crosstab(x, useNA=TRUE, long=TRUE)

[Package raster version 3.5-2 Index]