cut {raster}	R Documentation

Convert values to classes

Description

Cut uses the base function cut to classify the values of a Raster* object according to which interval they fall in. The intervals are defined by the argument breaks. The leftmost interval corresponds to level one, the next leftmost to level two and so on.

Usage


cut(x, ...) 

Arguments

x
A Raster* object

...
additional arguments. See cut

Value

Raster* object

See Also

subs, reclassify, calc

Examples



r <- raster(ncols=36, nrows=18)
values(r) <- rnorm(ncell(r)) 
breaks <- -2:2 * 3
rc <- cut(r, breaks=breaks)

[Package raster version 3.5-2 Index]