dim {raster}	R Documentation

Dimensions of a Raster* object

Description

Get or set the number of rows, columns, and layers of a Raster* object. You cannot use this function to set the dimensions of a RasterStack object.

When setting the dimensions, you can provide a row number, or a vector with the row and the column number (for a RasterLayer and a RasterBrick), or a row and column number and the number of layers (only for a RasterBrick)

Usage


## S4 method for signature 'BasicRaster'
dim(x)

Arguments

x
Raster(* object

Value

Integer or Raster* object

See Also

ncell, extent, res

Examples


r <- raster()
dim(r)
dim(r) <- c(18) 
dim(r)
dim(r) <- c(18, 36) 
dim(r)
b <- brick(r)
dim(b)
dim(b) <- c(10, 10, 5)
dim(b)

[Package raster version 3.5-2 Index]