draw {raster}	R Documentation

Draw a line or polygon

Description

Draw a line or polygon on a plot (map) and save it for later use. After calling the function, start clicking on the map. To finish, right-click and select 'stop'.

Usage


drawPoly(sp=TRUE, col='red', lwd=2, ...)
drawLine(sp=TRUE, col='red', lwd=2, ...)

Arguments

sp
logical. If TRUE, the output will be a sp object (SpatialPolygons or SpatialLines). Otherwise a matrix of coordinates is returned

col
the color of the lines to be drawn

lwd
the width of the lines to be drawn

...
additional arguments padded to locator

Value

If sp==TRUE a SpatialPolygons or SpatialLines object; otherwise a matrix of coordinates

See Also

locator

[Package raster version 3.5-2 Index]