drawExtent {raster}	R Documentation

Create an Extent object by drawing on a map

Description

Click on two points of a plot (map) to obtain an object of class Extent ('bounding box')

Usage


drawExtent(show=TRUE, col="red") 

Arguments

show
logical. If TRUE, the extent will be drawn on the map

col
sets the color of the lines of the extent

Value

Extent

Examples


## Not run: 
r1 <- raster(nrow=10, ncol=10)
values(r1) <- runif(ncell(r1))
plot(r1)
# after running the following line, click on the map twice
e <- drawExtent()
# after running the following line, click on the map twice
mean(values(crop(r1, drawExtent())))

## End(Not run)

[Package raster version 3.5-2 Index]