extremeValues {raster}	R Documentation

Minimum and maximum values

Description

Returns the minimum or maximum value of a RasterLayer or layer in a RasterStack

Usage


minValue(x, ...)
maxValue(x, ...)

Arguments

x
RasterLayer or RasterStack object

...
Additional argument: layer number (for RasterStack or RasterBrick objects)

Details

If a Raster* object is created from a file on disk, the min and max values are often not known (depending on the file format). You can use setMinMax to set them in the Raster* object.

Value

a number

Examples


r <- raster()
r <- setValues(r, 1:ncell(r))
minValue(r)
maxValue(r)
r <- setValues(r, round(100 * runif(ncell(r)) + 0.5))
minValue(r)
maxValue(r)

r <- raster(system.file("external/test.grd", package="raster"))
minValue(r)
maxValue(r)

[Package raster version 3.5-2 Index]