filename {raster}	R Documentation

Filename

Description

Get the filename of a Raster* object. You cannot set the filename of an object (except for RasterStack objects); but you can provide a 'filename= ' argument to a function that creates a new RasterLayer or RasterBrick* object.

Usage


filename(x)

Arguments

x
A Raster* object

Value

a Raster* object

Examples

 
r <- raster( system.file("external/test.grd", package="raster") )
filename(r)

[Package raster version 3.5-2 Index]