Get the coordinates of a vector type Spatial* object

Extract the coordinates of a Spatial object

# S4 method for SpatialPolygons
geom(x, sepNA=FALSE, ...)
# S4 method for SpatialLines
geom(x, sepNA=FALSE, ...)
# S4 method for SpatialPoints
geom(x, ...)
# S4 method for data.frame
geom(x, d, gt, crs, ...)

Arguments

x
SpatialPolygons*, SpatialLines*, or SpatialPoints* object; or a data.frame

sepNA
logical. If TRUE, geometries are separated by a row with NA values

...
additional arguments, none implemented

d
data.frame that matches the number of objects in data.frame x

gt
character. geometry type. Must be one of "polygons", "lines", "points"

crs
character. PROJ.4 crs string

Value

Matrix with 6, (5 SpatialLines), or 3 (SpatialPoints) columns. object (sequential object number) part (sequential part number within the object; not for SpatialPoints), cump (cumulative part number; not for SpatialPoints), hole (is this a hole or not; only for SpatialPolygons), x (x coordinate or longitude), y (y coordinate or latitude)

See also

Examples

p <- readRDS(system.file("external/lux.rds", package="raster"))
x <- geom(p)
head(x)

#>      object part cump hole        x        y
#> [1,]      1    1    1    0 6.026519 50.17767
#> [2,]      1    1    1    0 6.031361 50.16563
#> [3,]      1    1    1    0 6.035646 50.16410
#> [4,]      1    1    1    0 6.042747 50.16157
#> [5,]      1    1    1    0 6.043894 50.16116
#> [6,]      1    1    1    0 6.048243 50.16008

  
# and back to a SpatialPolygonsDataFrame  
x <- data.frame(x)
sp <- as(x, "SpatialPolygons")
crs(sp) <- crs(p)
spdf <- SpatialPolygonsDataFrame(sp, data.frame(p), match.ID=FALSE)