Get geographic data

Get geographic data for anywhere in the world. Data are read from files that are first downloaded if necessary. Function ccodes returns country names and the ISO codes

See the geodata package for access to more recent versions of the data available through this function, and for many other datasets.

getData(name, download=TRUE, path="", ...)
ccodes()

Arguments

name

Data set name, currently supported are 'GADM', 'countries', 'SRTM', 'alt', and 'worldclim'. See Details for more info

download

Logical. If TRUE data will be downloaded if not locally available

path

Character. Path name indicating where to store the data. Default is the current working directory

...

Additional required (!) parameters. These are data set specific. See Details

Value

A spatial object (Raster* or Spatial*)

Details

'alt' stands for altitude (elevation); the data were aggregated from SRTM 90 m resolution data between -60 and 60 latitude. 'GADM' is a database of global administrative boundaries. 'worldclim' is a database of global interpolated climate data. 'SRTM' refers to the hole-filled CGIAR-SRTM (90 m resolution). 'countries' has polygons for all countries at a higher resolution than the 'wrld_simpl' data in the maptools package .

If name is 'alt' or 'GADM' you must provide a 'country=' argument. Countries are specified by their 3 letter ISO codes. Use getData('ISO3') to see these codes. In the case of GADM you must also provide the level of administrative subdivision (0=country, 1=first level subdivision). In the case of alt you can set 'mask' to FALSE. If it is TRUE values for neighbouring countries are set to NA. For example:

getData('GADM', country='FRA', level=1)

getData('alt', country='FRA', mask=TRUE)

If name is 'SRTM' you must provide 'lon' and 'lat' arguments (longitude and latitude). These should be single numbers somewhere within the SRTM tile that you want.

getData('SRTM', lon=5, lat=45)

If name='worldclim' you must also provide arguments var, and a resolution res. Valid variables names are 'tmin', 'tmax', 'prec' and 'bio'. Valid resolutions are 0.5, 2.5, 5, and 10 (minutes of a degree). In the case of res=0.5, you must also provide a lon and lat argument for a tile; for the lower resolutions global data will be downloaded. In all cases there are 12 (monthly) files for each variable except for 'bio' which contains 19 files.

getData('worldclim', var='tmin', res=0.5, lon=5, lat=45)

getData('worldclim', var='bio', res=10)

To get (projected) future climate data (CMIP5), you must provide arguments var and res as above. Only resolutions 2.5, 5, and 10 are currently available. In addition, you need to provide model, rcp and year. For example,

getData('CMIP5', var='tmin', res=10, rcp=85, model='AC', year=70)

function (var, model, rcp, year, res, lon, lat, path, download = TRUE)

'model' should be one of "AC", "BC", "CC", "CE", "CN", "GF", "GD", "GS", "HD", "HG", "HE", "IN", "IP", "MI", "MR", "MC", "MP", "MG", or "NO".

'rcp' should be one of 26, 45, 60, or 85.

'year' should be 50 or 70

Not all combinations are available. See www.worldclim.org for details.

References

https://www.worldclim.org

https://gadm.org

https://srtm.csi.cgiar.org/

https://diva-gis.org/gdata