hdr {raster}	R Documentation

Header files

Description

Write header files to use together with raster binary files to read the data in other applications.

Usage


hdr(x, format, extension='.wld', filename='') 

Arguments

x
RasterLayer or RasterBrick object associated with a binary values file on disk

format
Type of header file: 'VRT', 'BIL', 'ENVI', 'ErdasRaw', 'IDRISI', 'SAGA', 'RASTER', 'WORLDFILE', 'PRJ'

extension
File extension, only used with an ESRI worldfile (format='WORLDFILE')

filename
character. Need to be provided if x is not associated with a file

Details

The RasterLayer object must be associated with a file on disk.

You can use writeRaster to save a existing file in another format. But if you have a file in a 'raster' format (or similar), you can also only export a header file, and use the data file (.gri) that already exists. The function can write a VRT (GDAL virtual raster) header (.vrt); an ENVI or BIL header (.hdr) file; an Erdas Raw (.raw) header file; an IDRISI (.rdc) or SAGA (.sgrd). This (hopefully) allows for reading the binary data (.gri), perhaps after changing the file extension, in other programs such as ENVI or ArcGIS.

See Also

writeRaster, writeGDAL

Examples

 
## Not run: 
r <- raster(system.file("external/test.grd", package="raster"))
r <- writeRaster(r, filename='export.grd', overwrite=TRUE)
hdr(r, format="ENVI") 

## End(Not run)

[Package raster version 3.5-2 Index]