hillShade {raster}	R Documentation

Hill shading

Description

Compute hill shade from slope and aspect layers (both in radians). Slope and aspect can be computed with function terrain.

A hill shade layer is often used as a backdrop on top of which another, semi-transparent, layer is drawn.

Usage


hillShade(slope, aspect, angle=45, direction=0, filename='', normalize=FALSE, ...)

Arguments

slope
RasterLayer object with slope values (in radians)

aspect
RasterLayer object with aspect values (in radians)

angle
The the elevation angle of the light source (sun), in degrees

direction
The direction (azimuth) angle of the light source (sun), in degrees

filename
Character. Optional filename

normalize
Logical. If TRUE, values below zero are set to zero and the results are multiplied with 255

...
Standard additional arguments for writing RasterLayer files

Author(s)

Andrew Bevan, Robert J. Hijmans

References

Horn, B.K.P., 1981. Hill shading and the reflectance map. Proceedings of the IEEE 69(1):14-47

See Also

terrain

Examples


## Not run: 
alt <- getData('alt', country='CHE')
slope <- terrain(alt, opt='slope')
aspect <- terrain(alt, opt='aspect')
hill <- hillShade(slope, aspect, 40, 270)
plot(hill, col=grey(0:100/100), legend=FALSE, main='Switzerland')
plot(alt, col=rainbow(25, alpha=0.35), add=TRUE)

## End(Not run)

[Package raster version 3.5-2 Index]