hist {raster}	R Documentation

Histogram

Description

Create a histogram of the values of a RasterLayer. For large datasets a sample is used.

Usage


## S4 method for signature 'Raster'
hist(x, layer, maxpixels=100000, plot=TRUE, main, ...)

Arguments

x
Raster* object

layer
integer (or character) to indicate layer number (or name). Can be used to subset the layers to plot in a multilayer Raster* object

maxpixels
integer. To regularly subsample very large objects

plot
logical. Plot the histogram or only return the histogram values

main
character. Main title(s) for the plot. Default is the value of names

...
Additional arguments. See under Methods and at hist

Value

This function is principally used for the side-effect of plotting a histogram, but it also returns an S3 object of class 'histogram' (invisibly if plot=TRUE).

See Also

pairs, boxplot

Examples

 
r1 <- raster(nrows=50, ncols=50)
r1 <- setValues(r1, runif(ncell(r1)))
r2 <- setValues(r1, runif(ncell(r1)))
rs <- r1 + r2
rp <- r1 * r2
par(mfrow=c(2,2))
plot(rs, main='sum')
plot(rp, main='product')
hist(rs)
a = hist(rp)
a

[Package raster version 3.5-2 Index]