inifile {raster}	R Documentation

Read a .ini file

Description

This function reads '.ini' files. These are text file databases that are organized in [sections] containing pairs of "name = value".

Usage


readIniFile(filename, token='=', commenttoken=';', aslist=FALSE, case)

Arguments

filename
Character. Filename of the .ini file

token
Character. The character that separates the "name" (variable name) from the "value"

commenttoken
Character. This token and everything that follows on the same line is considered a 'comment' that is not for machine consumption and is ignored in processing

aslist
Logical. Should the values be returned as a list

case
Optional. Function that operates on the text, such as toupper or tolower

Details

This function allows for using inistrings that have "=" as part of a value (but the token cannot be part of the 'name' of a variable!). Sections can be missing.

Value

A n*3 matrix of characters with columns: section, name, value; or a list if aslist=TRUE.

[Package raster version 3.5-2 Index]