modal value

Compute the mode for a vector of numbers, or across raster layers. The mode, or modal value, is the most frequent value in a set of values.

# S4 method for class 'ANY'
modal(x, ..., ties='random', na.rm=FALSE, freq=FALSE)

# S4 method for class 'Raster'
modal(x, ..., ties='random', na.rm=FALSE, freq=FALSE)

Arguments

x

vector of numbers (typically integers), characters, logicals, or factors, or a Raster* object

...

additional argument of the same type as x

ties

character. Indicates how to treat ties. Either 'random', 'lowest', 'highest', 'first', or 'NA'

na.rm

logical. If TRUE, NA values are ignored. If FALSE, NA is returned if x has any NA values

freq

return the frequency of the modal value, instead of the modal value

Value

vector or RasterLayer. The vector has length 1 and is of the same type as x, except when x is a factor and additional arguments (values) are supplied, in which case the values are coerced to characters and a character value is returned.

Examples

data <- c(0,1,2,3,3,3,3,4,4,4,5,5,6,7,7,8,9,NA)
modal(data, na.rm=TRUE)
#> [1] 3