origin {raster}	R Documentation

Origin

Description

Origin returns (or sets) the coordinates of the point of origin of a Raster* object. This is the point closest to (0, 0) that you could get if you moved towards that point in steps of the x and y resolution.

Usage


origin(x, ...)
origin(x) <- value

Arguments

x
Raster* object

value
numeric vector of length 1 or 2

...
additional arguments. None implemented

Value

A vector of two numbers (x and y coordinates), or a changed origin for x.

See Also

extent

Examples


r <- raster(xmn=-0.5, xmx = 9.5, ncols=10)
origin(r)
r
origin(r) <- 0
r

[Package raster version 3.5-2 Index]