persp {raster}	R Documentation

Perspective plot

Description

Perspective plot of a RasterLayer. This is an implementation of a generic function in the graphics package.

Usage


## S4 method for signature 'RasterLayer'
persp(x,  maxpixels=1e+05, ext=NULL, ...)

## S4 method for signature 'RasterStackBrick'
persp(x, y=1, maxpixels=10000, ext=NULL, ...)

Arguments

x
Raster* object

y
integer > 0 & <= nlayers(x) to select the layer of x if x is a RasterLayer or RasterBrick

maxpixels
integer > 0. Maximum number of cells to use for the plot. If maxpixels < ncell(x), sampleRegular is used before plotting

ext
Extent. Can be used to zoom in to a region (see also zoom and crop(x, drawExtent())

...
Any argument that can be passed to persp (graphics package)

See Also

plot3D, persp, contour, plot

Examples

 
r <- raster(system.file("external/test.grd", package="raster"))
persp(r)

[Package raster version 3.5-2 Index]