properties {raster}	R Documentation

Raster file properties

Description

Properties of the values of the file that a RasterLayer object points to

dataSize returns the number of bytes used for each value (pixel, grid cell) dataSigned is TRUE for data types that include negative numbers.

Usage


dataSize(object)
dataSigned(object)

Arguments

object
Raster* object

Value

varies

See Also

filename

Examples


r <- raster(system.file("external/test.grd", package="raster"))
dataSize(r)
dataSigned(r)
dataType(r)

[Package raster version 3.5-2 Index]