quantile {raster}	R Documentation

Raster quantiles

Description

Compute quantiles for the cell values of a RasterLayer. If you want to compute quantiles for each cell across a number of layers, you can use calc(x, fun=quantile).

Usage


quantile(x, ...) 

Arguments

x
Raster object

...
Additional arguments: na.rm=TRUE, ncells=NULL, and additional arguments to the stats::quantile function, see quantile

ncells can be used to set the number of cells to be sampled, for very large raster datasets.

Value

A vector of quantiles

See Also

density, cellStats

Examples


r <- raster(ncol=100, nrow=100)
values(r) <- rnorm(ncell(r), 0, 50)
quantile(r)
quantile(r, probs = c(0.25, 0.75), type=7,names = FALSE)

[Package raster version 3.5-2 Index]