Global options for the raster package

Set, inspect, reset, save a number of global options used by the raster package.

Most of these options are used when writing files to disk. They can be ignored by specific functions if the corresponding argument is provided as an argument to these functions.

The default location is returned by rasterTmpDir. It is the same as that of the R temp directory but you can change it (for the current session) with rasterOptions(tmpdir="path").

To permanently set any of these options, you can add them to <your R installation>/etc/Rprofile.site>. For example, to change the default directory used to save temporary files, add a line like this: options(rasterTmpDir='c:/temp/') to that file. All temporary raster files in that folder that are older than 24 hrs are deleted when the raster package is loaded.

Function tmpDir returns the location of the temporary files

rasterOptions(format, overwrite, datatype, tmpdir, tmptime, progress,
     timer, chunksize, minmemory, maxmemory, memfrac, todisk, setfileext, 
   tolerance, standardnames, depracatedwarnings, addheader, default=FALSE)

tmpDir(create=TRUE)

Arguments

format

character. The default file format to use. See writeFormats

overwrite

logical. The default value for overwriting existing files. If TRUE, existing files will be overwritten

datatype

character. The default data type to use. See dataType

tmpdir

character. The default location for writing temporary files; See rasterTmpFile

tmptime

number > 1. The number of hours after which a temporary file will be deleted. As files are deleted when loading the raster package, this option is only useful if you save this option so that it is loaded when starting a new session

progress

character. Valid values are "text", "window" and "" (the default in most functions, no progress bar)

timer

Logical. If TRUE, the time it took to complete the function is printed

chunksize

integer. Maximum number of bytes to read/write in a single chunk while processing (chunk by chunk) disk based Raster* objects

maxmemory

numeric. Maximum number of bytes to read into memory. If a process is expected to require more than this value, canProcessInMemory will return FALSE. It cannot be set to a value smaller than 10000

minmemory

numeric. Minimum number of bytes that are guaranteed to be fit into memory. If a process is expected to require more than this value, RAM available will be estimated. It cannot be set to a value smaller than 10000

memfrac

numeric. Fraction of available RAM that may be used by a process

todisk

logical. For debugging only. Default is FALSE and should normally not be changed. If TRUE, results are always written to disk, even if no filename is supplied (a temporary filename is used)

setfileext

logical. Default is TRUE. If TRUE, the file extension will be changed when writing (if known for the file type). E.g. GTiff files will be saved with the .tif extension

tolerance

numeric. The tolerance used when comparing the origin and resolution of Raster* objects. Expressed as the fraction of a single cell. This should be a number between 0 and 0.5

standardnames

logical. Default is TRUE. Should names be standardized to be syntactically valid names (using make.names)

depracatedwarnings

logical. If TRUE (the default) a warning is generated when a depracated (obsolete) function is used

addheader

character. If not equal to '' (the default) an additional header file is written when a raster format file (grd/gri) is written. Supported formats are as in hdr

default

logical. If TRUE, all options are set to their default values

create

logical. If TRUE, the temporary files directory is created if it does not exist

Value

list of the current options (invisibly). If no arguments are provided the options are printed.

See also

Examples

if (FALSE) {
rasterOptions()
rasterOptions(chunksize=2e+07)
}