readAll {raster}	R Documentation

Read values from disk

Description

Read all values from a raster file associated with a Raster* object into memory. This function should normally not be used. In most cases getValues or getValuesBlock is more appropriate as readAll will fail when there is no file associated with the RasterLayer (values may only exist in memory).

Usage


readAll(object)

Arguments

object
a Raster* object

See Also

getValues, getValuesBlock, extract

Examples


r <- raster(system.file("external/test.grd", package="raster"))
r <- readAll(r)

[Package raster version 3.5-2 Index]