rectify {raster}	R Documentation

rectify a Raster object

Description

rectify changes a rotated Raster* object into a non-rotated (rectangular) object. This is wrapper function around resample.

Usage


## S4 method for signature 'Raster'
rectify(x, ext, res, method='ngb', filename='', ...) 

Arguments

x
Raster* object to be rectified

ext
Optional. Extent object or object from which an Extent object can be extracted

res
Optional. Single or two numbers to set the resolution

method
Method used to compute values for the new RasterLayer, should be "bilinear" for bilinear interpolation, or "ngb" for nearest neighbor

filename
Character. Output filename

...
Additional arguments as for writeRaster

Value

RasterLayer or RasterBrick object

[Package raster version 3.5-2 Index]