resample {raster}	R Documentation

Resample a Raster object

Description

Resample transfers values between non matching Raster* objects (in terms of origin and resolution). Use projectRaster if the target has a different coordinate reference system (projection).

Before using resample, you may want to consider using these other functions instead: aggregate, disaggregate, crop, extend, merge.

Usage


## S4 method for signature 'Raster,Raster'
resample(x, y, method="bilinear", filename="", ...)

Arguments

x
Raster* object to be resampled

y
Raster* object with parameters that x should be resampled to

method
method used to compute values for the new RasterLayer, should be "bilinear" for bilinear interpolation, or "ngb" for using the nearest neighbor

filename
character. Output filename (optional)

...
Additional arguments as for writeRaster

Value

RasterLayer or RasterBrick object

Author(s)

Robert J. Hijmans and Joe Cheng

See Also

aggregate, disaggregate, crop, extend, merge, projectRaster

Examples


r <- raster(nrow=3, ncol=3)
values(r) <- 1:ncell(r)
s <- raster(nrow=10, ncol=10)
s <- resample(r, s, method='bilinear')
#par(mfrow=c(1,2))
#plot(r)
#plot(s)

[Package raster version 3.5-2 Index]