round {raster}	R Documentation

Integer values

Description

These functions take a single RasterLayer argument x and change its values to integers.

ceiling returns a RasterLayer with the smallest integers not less than the corresponding values of x.

floor returns a RasterLayer with the largest integers not greater than the corresponding values of x.

trunc returns a RasterLayer with the integers formed by truncating the values in x toward 0.

round returns a RasterLayer with values rounded to the specified number of digits (decimal places; default 0).

Details

see ?base::round

Value

a RasterLayer object

Methods

ceiling(x) floor(x) trunc(x, ...) round(x, digits = 0)

x

a RasterLayer object

digits

integer indicating the precision to be used

...

additional arguments

Examples


r <- raster(ncol=10, nrow=10)
values(r) <- runif(ncell(r)) * 10
s <- round(r)


[Package raster version 3.5-2 Index]