round Extent coordinates

use round(x, digits=0) to round the coordinates of an Extent object to the number of digits specified. This can be useful when dealing with a small imprecision in the data (e.g. 179.9999 instead of 180). floor and ceiling move the coordiantes to the outer or inner whole integer numbers.

It is also possible to use Arithmetic functions with Extent objects (but these work perhaps unexpectedly!)

See Math-methods for these (and many more) methods with Raster* objects.

# S4 method for Extent
floor(x)
# S4 method for Extent
ceiling(x)

Arguments

x
Extent object

See also

Examples

e <- extent(c(0.999999,  10.000011, -60.4, 60))
round(e)
#> class      : Extent 
#> xmin       : 1 
#> xmax       : 10 
#> ymin       : -60 
#> ymax       : 60 
ceiling(e)
#> class      : Extent 
#> xmin       : 1 
#> xmax       : 10 
#> ymin       : -60 
#> ymax       : 60 
floor(e)
#> class      : Extent 
#> xmin       : 0 
#> xmax       : 11 
#> ymin       : -61 
#> ymax       : 60