SampleInt {raster}	R Documentation

Sample integer values

Description

Take a random sample from a range of integer values between 1 and n. Its purpose is similar to that of sample, but that function fails when n is very large.

Usage


sampleInt(n, size, replace=FALSE)

Arguments

n
Positive number (integer); the number of items to choose from

size
Non-negative integer; the number of items to choose

replace
Logical. Should sampling be with replacement?

Value

vector of integer numbers

Examples


  sampleInt(1e+12, 10)
  
# this may fail:
#  sample.int(1e+12, 10)
#  sample.int(1e+9, 10)

[Package raster version 3.5-2 Index]