Random sample

Take a random sample from the cell values of a Raster* object (without replacement).

# S4 method for Raster
sampleRandom(x, size, na.rm=TRUE, ext=NULL, 
    cells=FALSE, rowcol=FALSE, xy=FALSE, sp=FALSE, asRaster=FALSE, ...)

Arguments

x
Raster* object

size
positive integer giving the number of items to choose

na.rm
logical. If TRUE (the default), NA values are removed from random sample

ext
Extent object. To limit regular sampling to the area within the extent

cells
logical. If TRUE, sampled cell numbers are also returned

rowcol
logical. If TRUE, sampled row and column numbers are also returned

xy
logical. If TRUE, coordinates of sampled cells are also returned

sp
logical. If TRUE, a SpatialPointsDataFrame is returned

asRaster
logical. If TRUE, a Raster* object is returned with random cells with values, all other cells with NA

...
Additional arguments as in writeRaster. Only relevant when asRaster=TRUE

Details

With argument na.rm=TRUE, the returned sample may be smaller than requested

Value

A vector, matrix (if cells=TRUE or x is a multi-layered object), or a SpatialPointsDataFrame (if sp=TRUE )

See also

Examples

 r <- raster(system.file("external/test.grd", package="raster"))
 sampleRandom(r, size=10)

#>  [1] 461.2969 318.2996 351.8303 392.7534 492.7690 243.2765 440.3402 345.6403
#>  [9] 515.7428 384.0551

 s <- stack(r, r)
 sampleRandom(s, size=5, cells=TRUE, sp=TRUE)

#> class       : SpatialPointsDataFrame 
#> features    : 5 
#> extent      : 178700, 180980, 329740, 333020  (xmin, xmax, ymin, ymax)
#> crs         : +proj=sterea +lat_0=52.1561605555556 +lon_0=5.38763888888889 +k=0.9999079 +x_0=155000 +y_0=463000 +datum=WGS84 +units=m +no_defs 
#> variables   : 3
#> names       : cell,           test.1,           test.2 
#> min values  : 1984, 245.001937866211, 245.001937866211 
#> max values  : 8502, 433.913879394531, 433.913879394531