Regular sample

Take a systematic sample from a Raster* object.

# S4 method for class 'Raster'
sampleRegular(x, size, ext=NULL, cells=FALSE, xy=FALSE, asRaster=FALSE, 
            sp=FALSE, ...)

Arguments

x

Raster object

size

positive integer giving the number of items to choose.

ext

Extent. To limit regular sampling to the area within that box

cells

logical. Also return sampled cell numbers (if asRaster=FALSE)

xy

logical. If TRUE, coordinates of sampled cells are also returned

asRaster

logical. If TRUE, a RasterLayer or RasterBrick is returned, rather than the sampled values

sp

logical. If TRUE, a SpatialPointsDataFrame is returned

...

additional arguments. None implemented

Value

A vector (single layer object), matrix (multi-layered object; or if cells=TRUE, or xy=TRUE), Raster* object (if asRaster=TRUE), or SpatialPointsDataFrame (if sp=TRUE)

See also

Examples

 r <- raster(system.file("external/test.grd", package="raster"))
 v <- sampleRegular(r, size=100)
 x <- sampleRegular(r, size=100, asRaster=TRUE)