stackSave {raster}	R Documentation

Save or open a RasterStack file

Description

A RasterStack is a collection of RasterLayers with the same spatial extent and resolution. They can be created from RasterLayer objects, or from file names. These two functions allow you to save the references to raster files and recreate a rasterStack object later. They only work if the RasterStack points to layers that have their values on disk. The values are not saved, only the references to the files.

Usage


stackOpen(stackfile)
stackSave(x, filename)

Arguments

stackfile
Filename for the RasterStack (to save it on disk)

x
RasterStack object

filename
File name

Details

When a RasterStack is saved to a file, only pointers (filenames) to raster datasets are saved, not the data. If the name or location of a raster file changes, the RasterStack becomes invalid.

Value

RasterStack object

See Also

writeRaster, stack, addLayer

Examples


file <- system.file("external/test.grd", package="raster")
s <- stack(c(file, file))

## Not run: 
s <- stackSave(s, "mystack")
# note that filename adds an extension .stk to a stackfile  
s2 <- stackOpen("mystack.stk")
s2

## End(Not run)

[Package raster version 3.5-2 Index]