scale {raster}	R Documentation

Scale values

Description

Center and/or scale raster data

Usage


## S4 method for signature 'Raster'
scale(x, center=TRUE, scale=TRUE)

Arguments

x
Raster* object

center
logical or numeric. If TRUE, centering is done by subtracting the layer means (omitting NAs), and if FALSE, no centering is done. If center is a numeric vector with length equal to the nlayers(x), then each layer of x has the corresponding value from center subtracted from it.

scale
logical or numeric. If TRUE, scaling is done by dividing the (centered) layers of x by their standard deviations if center is TRUE, and the root mean square otherwise. If scale is FALSE, no scaling is done. If scale is a numeric vector with length equal to nlayers(x), each layer of x is divided by the corresponding value. Scaling is done after centering.

Value

Raster* object

See Also

scale

Examples


b <- brick(system.file("external/rlogo.grd", package="raster"))
bs <- scale(b)

[Package raster version 3.5-2 Index]