setValues {raster}	R Documentation

Set values of a Raster object

Description

Assign (new) values to a Raster* object.

Usage


## S4 method for signature 'RasterLayer'
setValues(x, values, ...)

## S4 method for signature 'RasterBrick'
setValues(x, values, layer=-1, ...)

## S4 method for signature 'RasterStack'
setValues(x, values, layer=-1, ...)

## S4 method for signature 'RasterLayerSparse'
setValues(x, values, index=NULL, ...)

values(x) <- value

Arguments

x
A Raster*

values
Cell values to associate with the Raster* object. There should be values for all cells

value
Cell values to associate with the Raster* object. There should be values for all cells

layer
Layer number (only relevant for RasterBrick and RasterStack objects). If missing, the values of all layers is set

index
Cell numbers corresponding to the values

...
Additional arguments (none implemented)

Value

a Raster* object

Note

While you can access the 'values' slot of the objects directly, you would do that at your own peril because when setting values, multiple slots need to be changed; which is what setValues takes care of.

See Also

replacement

Examples

 
r <- raster(ncol=10, nrow=10)
vals <- 1:ncell(r)
r <- setValues(r, vals)
# equivalent to
values(r) <- vals

[Package raster version 3.5-2 Index]