Slope and aspect {raster}	R Documentation

Slope and aspect

Description

This is a deprecated function. Use terrain instead.

Usage

slopeAspect(dem, filename='', out=c('slope', 'aspect'), unit='radians', 
                 neighbors=8, flatAspect, ...)  

Arguments

dem
RasterLayer object with elevation values in map units, or in meters when the crs is longitude/latitude

filename
Character. Filename. optional

out
Character vector containing one or more of these options: 'slope', 'aspect'

unit
Character. 'degrees' or 'radians'

neighbors
Integer. Indicating how many neighboring cells to use to compute slope for any cell. Either 8 (queen case) or 4 (rook case), see Details

flatAspect
Numeric or NA. What value to use for aspect when slope is zero (and hence the aspect is undefined)? The default value is 90 degrees (or 0.5*pi radians)

...
Standard additional arguments for writing RasterLayer files

See Also

terrain

[Package raster version 2.7-15 Index]