Summary {raster}	R Documentation

Summary

Description

Summarize a Raster* object. A sample is used for very large files.

Usage


## S4 method for signature 'RasterLayer'
summary(object, maxsamp=100000, ...)

Arguments

object
Raster* object

maxsamp
positive integer. Sample size used for large datasets

...
additional arguments. None implemented

Value

matrix with (an estimate of) the median, minimum and maximum values, the first and third quartiles, and the number of cells with NA values

See Also

cellStats, link[raster]{quantile}

[Package raster version 3.5-2 Index]