unique {raster}	R Documentation

Unique values

Description

This function returns the unique values in a RasterLayer object or the unique combinations of the layers in a multilayer object.

Usage


## S4 method for signature 'RasterLayer,missing'
unique(x, incomparables=FALSE, na.last=NA, progress="", ...) 

## S4 method for signature 'RasterStackBrick,missing'
unique(x, incomparables=FALSE, na.last=NA, progress="", ...) 

Arguments

x
Raster object

incomparables
must be missing. The default value FALSE is used. See unique

na.last
logical. for controlling the treatment of NAs. If TRUE, missing values in the data are put last; if FALSE, they are put first; if NA, they are removed.

progress
character. Use "text" or "window" for a progress indicator

...
additional arguments. as in unique

Value

vector or matrix

See Also

unique

Examples


r <- raster(ncol=10, nrow=10)
values(r) <- round(runif(ncell(r))*10)
unique(r)
unique(stack(r, round(r/2)))

[Package raster version 3.5-2 Index]