validCell {raster}	R Documentation

Validity of a cell, column or row number

Description

Simple helper functions to determine if a row, column or cell number is valid for a certain Raster* object

Usage


validCell(object, cell) 
validCol(object, colnr) 
validRow(object, rownr) 

Arguments

object
Raster* object (or a SpatialPixels* or SpatialGrid* object)

cell
cell number(s)

colnr
column number; or vector of column numbers

rownr
row number; or vector of row numbers

Value

logical value

Examples


#using a new default raster (1 degree global)
r <- raster()
validCell(r, c(-1, 0, 1))
validRow(r, c(-1, 1, 100, 10000))

[Package raster version 3.5-2 Index]