File types for writing

List supported file types for writing RasterLayer values to disk.

When a function writes a file to disk, the file format is determined by the 'format=' argument if supplied, or else by the file extension (if the extension is known). If other cases the default format is used. The 'factory-fresh' default format is 'raster', but this can be changed using rasterOptions.

writeFormats()

Details

writeFormats returns a matrix of the file formats (the "drivers") that are supported.

Supported formats include:

File type	Long name	default extension	Multiband support

raster	'Native' raster package format	.grd	Yes

ascii	ESRI Ascii	.asc	No

SAGA	SAGA GIS	.sdat	No

IDRISI	IDRISI	.rst	No

CDF	netCDF (requires ncdf4)	.nc	Yes

GTiff	GeoTiff (requires rgdal)	.tif	Yes

ENVI	ENVI .hdr Labelled	.envi	Yes

EHdr	ESRI .hdr Labelled	.bil	Yes

HFA	Erdas Imagine Images (.img)	.img	Yes

See also

Examples

writeFormats() 

#>       name        long_name                                
#>  [1,] "raster"    "R-raster"                               
#>  [2,] "SAGA"      "SAGA GIS"                               
#>  [3,] "IDRISI"    "IDRISI"                                 
#>  [4,] "IDRISIold" "IDRISI (img/doc)"                       
#>  [5,] "BIL"       "Band by Line"                           
#>  [6,] "BSQ"       "Band Sequential"                        
#>  [7,] "BIP"       "Band by Pixel"                          
#>  [8,] "ascii"     "Arc ASCII"                              
#>  [9,] "CDF"       "NetCDF"                                 
#> [10,] "ADRG"      "ARC Digitized Raster Graphics"          
#> [11,] "BAG"       "Bathymetry Attributed Grid"             
#> [12,] "BMP"       "MS Windows Device Independent Bitmap"   
#> [13,] "BT"        "VTP .bt (Binary Terrain) 1.3 Format"    
#> [14,] "BYN"       "Natural Resources Canada's Geoid"       
#> [15,] "CTable2"   "CTable2 Datum Grid Shift"               
#> [16,] "EHdr"      "ESRI .hdr Labelled"                     
#> [17,] "ELAS"      "ELAS"                                   
#> [18,] "ENVI"      "ENVI .hdr Labelled"                     
#> [19,] "ERS"       "ERMapper .ers Labelled"                 
#> [20,] "GPKG"      "GeoPackage"                             
#> [21,] "GS7BG"     "Golden Software 7 Binary Grid (.grd)"   
#> [22,] "GSBG"      "Golden Software Binary Grid (.grd)"     
#> [23,] "GTiff"     "GeoTIFF"                                
#> [24,] "GTX"       "NOAA Vertical Datum .GTX"               
#> [25,] "HDF4Image" "HDF4 Dataset"                           
#> [26,] "HFA"       "Erdas Imagine Images (.img)"            
#> [27,] "IDA"       "Image Data and Analysis"                
#> [28,] "ILWIS"     "ILWIS Raster Map"                       
#> [29,] "INGR"      "Intergraph Raster"                      
#> [30,] "ISCE"      "ISCE raster"                            
#> [31,] "ISIS2"     "USGS Astrogeology ISIS cube (Version 2)"
#> [32,] "ISIS3"     "USGS Astrogeology ISIS cube (Version 3)"
#> [33,] "KRO"       "KOLOR Raw"                              
#> [34,] "LAN"       "Erdas .LAN/.GIS"                        
#> [35,] "Leveller"  "Leveller heightfield"                   
#> [36,] "MBTiles"   "MBTiles"                                
#> [37,] "MRF"       "Meta Raster Format"                     
#> [38,] "netCDF"    "Network Common Data Format"             
#> [39,] "NGW"       "NextGIS Web"                            
#> [40,] "NITF"      "National Imagery Transmission Format"   
#> [41,] "NTv2"      "NTv2 Datum Grid Shift"                  
#> [42,] "NWT_GRD"   "Northwood Numeric Grid Format .grd/.tab"
#> [43,] "PAux"      "PCI .aux Labelled"                      
#> [44,] "PCIDSK"    "PCIDSK Database File"                   
#> [45,] "PCRaster"  "PCRaster Raster File"                   
#> [46,] "PDF"       "Geospatial PDF"                         
#> [47,] "PDS4"      "NASA Planetary Data System 4"           
#> [48,] "PNM"       "Portable Pixmap Format (netpbm)"        
#> [49,] "RMF"       "Raster Matrix Format"                   
#> [50,] "ROI_PAC"   "ROI_PAC raster"                         
#> [51,] "RRASTER"   "R Raster"                               
#> [52,] "RST"       "Idrisi Raster A.1"                      
#> [53,] "SAGA"      "SAGA GIS Binary Grid (.sdat, .sg-grd-z)"
#> [54,] "SGI"       "SGI Image File Format 1.0"              
#> [55,] "Terragen"  "Terragen heightfield"                   
#> [56,] "VICAR"     "MIPL VICAR file"