zoom {raster}	R Documentation

Zoom in on a map

Description

Zoom in on a map (plot) by providing a new extent, by default this is done by clicking twice on the map.

Usage


zoom(x, ...) 
## S4 method for signature 'Raster'
zoom(x, ext=drawExtent(), maxpixels=100000, layer=1, new=TRUE, useRaster=TRUE, ...)

## S4 method for signature 'Spatial'
zoom(x, ext=drawExtent(), new=TRUE, ...)

## S4 method for signature 'missing'
zoom(x, ext=drawExtent(), new=TRUE, ...)

Arguments

x
Raster* or Spatial* (vector type) object

ext
Extent object, or other object from which an extent can be extracted

maxpixels
Maximum number of pixels used for the map

layer
Positive integer to select the layer to be used if x is a mutilayer Raster object

new
Logical. If TRUE, the zoomed in map will appear on a new device (window)

useRaster
Logical. If TRUE, a bitmap raster is used to plot the image instead of polygons

...
additional paramters for base plot

Value

Extent object (invisibly)

See Also

drawExtent, plot

[Package raster version 3.5-2 Index]