boxplot {raster}	R Documentation

Box plot of Raster objects

Description

Box plot of layers in a Raster object

Usage

## S4 method for signature 'RasterStackBrick'
boxplot(x, maxpixels=100000, ...)

## S4 method for signature 'RasterLayer'
boxplot(x, y=NULL, maxpixels=100000, ...)

Arguments

x
Raster* object

y
If x is a RasterLayer object, y can be an additional RasterLayer to group the values of x by 'zone'

maxpixels
Integer. Number of pixels to sample from each layer of large Raster objects

...
Arguments passed to graphics::boxplot

See Also

pairs, hist

Examples

r1 <- r2 <- r3 <- raster(ncol=10, nrow=10)
r1[] <- rnorm(ncell(r1), 100, 40)
r2[] <- rnorm(ncell(r1), 80, 10)
r3[] <- rnorm(ncell(r1), 120, 30)
s <- stack(r1, r2, r3)
names(s) <- c('A', 'B', 'C')

boxplot(s, notch=TRUE, col=c('red', 'blue', 'orange'), main='Box plot', ylab='random' )

[Package raster version 2.7-15 Index]