calc {raster}	R Documentation

Calculate

Description

Calculate values for a new Raster* object from another Raster* object, using a formula.

If x is a RasterLayer, fun is typically a function that can take a single vector as input, and return a vector of values of the same length (e.g. sqrt). If x is a RasterStack or RasterBrick, fun should operate on a vector of values (one vector for each cell). calc returns a RasterLayer if fun returns a single value (e.g. sum) and it returns a RasterBrick if fun returns more than one number, e.g., fun=quantile.

In many cases, what can be achieved with calc, can also be accomplished with a more intuitive 'raster-algebra' notation (see Arith-methods). For example, r <- r * 2 instead of

r <- calc(r, fun=function(x){x * 2}, or r <- sum(s) instead of

r <- calc(s, fun=sum). However, calc should be faster when using complex formulas on large datasets. With calc it is possible to set an output filename and file type preferences.

See (overlay) to use functions that refer to specific layers, like (function(a,b,c){a + sqrt(b) / c})

Usage


## S4 method for signature 'Raster,function'
calc(x, fun, filename='', na.rm, forcefun=FALSE, forceapply=FALSE, ...)

Arguments

x
Raster* object

fun
function

filename
character. Output filename (optional)

na.rm
Remove NA values, if supported by 'fun' (only relevant when summarizing a multilayer Raster object into a RasterLayer)

forcefun
logical. Force calc to not use fun with apply; for use with ambiguous functions and for debugging (see Details)

forceapply
logical. Force calc to use fun with apply; for use with ambiguous functions and for debugging (see Details)

...
Additional arguments as for writeRaster

Details

The intent of some functions can be ambiguous. Consider:

library(raster)

r <- raster(volcano)

calc(r, function(x) x * 1:10)

In this case, the cell values are multiplied in a vectorized manner and a single layer is returned where the first cell has been multiplied with one, the second cell with two, the 11th cell with one again, and so on. But perhaps the intent was to create 10 new layers (x*1, x*2, ...)? This can be achieved by using argument forceapply=TRUE

calc(r, function(x) x * 1:10, forceapply=TRUE)

Value

a Raster* object

Note

For large objects calc will compute values chunk by chunk. This means that for the result of fun to be correct it should not depend on having access to _all_ values at once. For example, to scale the values of a Raster* object by subtracting its mean value (for each layer), you would _not_ do, for Raster object x:

calc(x, function(x)scale(x, scale=FALSE))

Because the mean value of each chunk will likely be different. Rather do something like

m <- cellStats(x, 'mean')

x - m

Author(s)

Robert J. Hijmans and Matteo Mattiuzzi

See Also

overlay , reclassify, Arith-methods, Math-methods

Examples


r <- raster(ncols=36, nrows=18)
values(r) <- 1:ncell(r)

# multiply values with 10
fun <- function(x) { x * 10 }
rc1 <- calc(r, fun)

# set values below 100 to NA. 
fun <- function(x) { x[x<100] <- NA; return(x) }
rc2 <- calc(r, fun)

# set NA values to -9999
fun <- function(x) { x[is.na(x)] <- -9999; return(x)} 
rc3 <- calc(rc2, fun)

# using a RasterStack as input
s <- stack(r, r*2, sqrt(r))
# return a RasterLayer
rs1 <- calc(s, sum)

# return a RasterBrick
rs2 <- calc(s, fun=function(x){x * 10})
# recycling by layer
rs3 <- calc(s, fun=function(x){x * c(1, 5, 10)})

# use overlay when you want to refer to individual layer in the function
# but it can be done with calc: 
rs4 <- calc(s, fun=function(x){x[1]+x[2]*x[3]})

## 
# Some regression examples
## 

# create data
r <- raster(nrow=10, ncol=10)
s1 <- lapply(1:12, function(i) setValues(r, rnorm(ncell(r), i, 3)))
s2 <- lapply(1:12, function(i) setValues(r, rnorm(ncell(r), i, 3)))
s1 <- stack(s1)
s2 <- stack(s2)

# regression of values in one brick (or stack) with another
s <- stack(s1, s2)
# s1 and s2 have 12 layers; coefficients[2] is the slope
fun <- function(x) { lm(x[1:12] ~ x[13:24])$coefficients[2] }
x1 <- calc(s, fun)

# regression of values in one brick (or stack) with 'time'
time <- 1:nlayers(s)
fun <- function(x) { lm(x ~ time)$coefficients[2] }
x2 <- calc(s, fun)

# get multiple layers, e.g. the slope _and_ intercept
fun <- function(x) { lm(x ~ time)$coefficients }
x3 <- calc(s, fun)


### A much (> 100 times) faster approach is to directly use 
### linear algebra and pre-compute some constants

## add 1 for a model with an intercept
X <- cbind(1, time)

## pre-computing constant part of least squares
invXtX <- solve(t(X) %*% X) %*% t(X)

## much reduced regression model; [2] is to get the slope
quickfun <- function(y) (invXtX %*% y)[2]
x4 <- calc(s, quickfun) 

[Package raster version 3.5-2 Index]