dataType {raster}	R Documentation

Data type

Description

Get the datatype of a RasterLayer object. The datatype determines the interpretation of values written to disk. Changing the datatype of a Raster* object does not directly affect the way they are stored in memory. For native file formats (.grd/.gri files) it does affect how values are read from file. This is not the case for file formats that are read via rgdal (such as .tif and .img files) or netcdf.

If you change the datatype of a RasterLayer and then read values from a native format file these may be completely wrong, so only do this for debugging or when the information in the header file was wrong. To set the datatype of a new file, you can give a 'datatype' argument to the functions that write values to disk (e.g. writeRaster).

Usage


dataType(x)
dataType(x) <- value

Arguments

x
A RasterLayer object

value
A data type (see below)

Details

Setting the data type is useful if you want to write values to disk. In other cases use functions such as round()

Datatypes are described by 5 characters. The first three indicate whether the values are integers, decimal number or logical values. The fourth character indicates the number of bytes used to save the values on disk, and the last character indicates whether the numbers are signed (i.e. can be negative and positive values) or not (only zero and positive values allowed)

The following datatypes are available:

Datatype definition 	 minimum possible value 	 maximum possible value

LOG1S 	 FALSE (0)	 TRUE (1)

INT1S 	 -127 	 127

INT1U 	 0 	 255

INT2S 	 -32,767	 32,767

INT2U 	 0 	 65,534

INT4S 	 -2,147,483,647 	 2,147,483,647

INT4U 	 0 	 4,294,967,296

FLT4S 	 -3.4e+38 	 3.4e+38

FLT8S 	 -1.7e+308 	 1.7e+308

For all integer types, except the single byte types, the lowest (signed) or highest (unsigned) value is used to store NA. Single byte files do not have NA values. Logical values are stored as signed single byte integers, they do have an NA value (-127)

INT4U is available but they are best avoided as R does not support 32-bit unsigned integers.

Value

Raster* object

Examples

 
r <- raster(system.file("external/test.grd", package="raster"))
dataType(r)
## Not run: 
s <- writeRaster(r, 'new.grd', datatype='INT2U', overwrite=TRUE)
dataType(s)

## End(Not run)

[Package raster version 3.5-2 Index]