datasource {raster}	R Documentation

Are values in memory and/or on disk?

Description

These are helper functions for programmers and for debugging that provide information about whether a Raster object has associated values, and if these are in memory or on disk.

fromDisk is TRUE if the data source is a file on disk; and FALSE if the object only exists in memory.

inMemoryi is TRUE if all values are currently in memory (RAM); and FALSE if not (in which case they either are on disk, or there are no values).

hasValues is TRUE if the object has cell values.

Usage


fromDisk(x)
## S4 method for signature 'BasicRaster'
inMemory(x)
## S4 method for signature 'BasicRaster'
hasValues(x)

Arguments

x
Raster* object

Value

Logical

Examples


rs <- raster(system.file("external/test.grd", package="raster"))
inMemory(rs)
fromDisk(rs)
rs <- readAll(rs)
inMemory(rs)
fromDisk(rs)
rs <- rs + 1
inMemory(rs)
fromDisk(rs)
rs <- raster(rs)
inMemory(rs)
fromDisk(rs)
rs <- setValues(rs, 1:ncell(rs))
inMemory(rs)
fromDisk(rs)
#rs <- writeRaster(rs, filename=rasterTmpFile(), overwrite=TRUE)
#inMemory(rs)
#fromDisk(rs)

[Package raster version 3.5-2 Index]