distanceFromPoints {raster}	R Documentation

Distance from points

Description

The function calculates the distance from a set of points to all cells of a RasterLayer.

The distance unit is in meters if the RasterLayer is not projected (+proj=longlat) and in map units (typically meters) when it is projected.

Usage

distanceFromPoints(object, xy, filename='', ...) 

Arguments

object
Raster object

xy
matrix of x and y coordinates, or a SpatialPoints* object.

filename
character. Optional filename for the output RasterLayer

...
Additional arguments as for writeRaster

Value

RasterLayer object

See Also

distance, gridDistance, pointDistance

Examples

r <- raster(ncol=36,nrow=18)
xy = c(0,0)
dist <- distanceFromPoints(r, xy) 
#plot(dist)

[Package raster version 2.5-8 Index]