Extract by index {raster}	R Documentation

Indexing to extract values of a Raster* object

Description

These are shorthand methods that call other methods that should normally be used, such as getValues, extract, crop.

object[i] can be used to access values of a Raster* object, using cell numbers. You can also use row and column numbers as index, using object[i,j] or object[i,] or object[,j]. In addition you can supply an Extent, SpatialPolygons, SpatialLines or SpatialPoints object.

If drop=TRUE (the default) cell values are returned (a vector for a RasterLayer, a matrix for a RasterStack or RasterBrick). If drop=FALSE a Raster* object is returned that has the extent covering the requested cells, and with all other non-requested cells within this extent set to NA.

If you supply a RasterLayer, its values will be used as logical (TRUE/FALSE) indices if both Raster objects have the same extent and resolution; otherwise the cell values within the extent of the RasterLayer are returned.

Double brackes '[[ ]]' can be used to extract one or more layers from a multi-layer object.

Methods

x[i]

x[i,j]

Arguments

x 	 a Raster* object

i 	 cell number(s), row number(s), a (logical) RasterLayer, Spatial* object

j 	 column number(s) (only available if i is (are) a row number(s))

drop 	 If TRUE, cell values are returned. Otherwise, a Raster* object is returned

See Also

getValues, setValues, extract, crop, rasterize

Examples


r <- raster(ncol=10, nrow=5)
values(r) <- 1:ncell(r) 

r[1]
r[1:10]
r[1,]
r[,1]
r[1:2, 1:2]

s <- stack(r, sqrt(r))
s[1:3]
s[[2]]


[Package raster version 3.5-2 Index]