freq {raster}	R Documentation

Frequency table

Description

Frequency table of the values of a RasterLayer.

Usage


## S4 method for signature 'RasterLayer'
freq(x, digits=0, value=NULL, useNA='ifany', progress='', ...)

## S4 method for signature 'RasterStackBrick'
freq(x, digits=0, value=NULL, useNA='ifany', merge=FALSE, progress='', ...)

Arguments

x
RasterLayer

digits
non-negative integer for rounding the cell values. Argument is passed to round

value
numeric, logical or NA. An optional single value to only count the number of cells with that value

useNA
character. What to do with NA values? Options are "no", "ifany", "always". See to table

progress
character to specify a progress bar. Choose from 'text', 'window', or ” (the default, no progress bar)

merge
logical. If TRUE the list will be merged into a single data.frame

...
additional arguments (none implemented)

Value

matrix (RasterLayer). List of matrices (one for each layer) or data.frame (if merge=TRUE) (RasterStack or RasterBrick)

See Also

crosstab and zonal

Examples


r <- raster(nrow=18, ncol=36)
values(r) <- runif(ncell(r))
r[1:5] <- NA
r <- r * r * r * 5
freq(r)

freq(r, value=2)

s <- stack(r, r*2, r*3)
freq(s, merge=TRUE)

[Package raster version 3.5-2 Index]