getValuesBlock {raster}	R Documentation

Get a block of raster cell values

Description

getValuesBlock returns values for a block (rectangular area) of values of a Raster* object.

Usage


## S4 method for signature 'RasterLayer'
getValuesBlock(x, row=1, nrows=1, col=1, ncols=(ncol(x)-col+1), format='', ...)

## S4 method for signature 'RasterBrick'
getValuesBlock(x, row=1, nrows=1, col=1, ncols=(ncol(x)-col+1), lyrs, ...)

## S4 method for signature 'RasterStack'
getValuesBlock(x, row=1, nrows=1, col=1, ncols=(ncol(x)-col+1), lyrs, ...)

Arguments

x
Raster* object

row
positive integer. Row number to start from, should be between 1 and nrow(x)

nrows
positive integer. How many rows? Default is 1

col
positive integer. Column number to start from, should be between 1 and ncol(x)

ncols
positive integer. How many columns? Default is the number of columns left after the start column

format
character. When x is a RasterLayer, if format='matrix' or format='m', a matrix is returned instead of a vector. If format='matrix', it is a nrow x ncol matrix. If format='m' it is a 1 column matrix (the benefit is that the type of output is now the same for all Raster objects)

lyrs
integer (vector). Which layers? Default is all layers (1:nlayers(x))

...
additional arguments (none implemented)

Value

matrix or vector (if (x=RasterLayer), unless format='matrix')

See Also

getValues

Examples


r <- raster(system.file("external/test.grd", package="raster"))
b <- getValuesBlock(r, row=100, nrows=3, col=10, ncols=5)
b 
b <- matrix(b, nrow=3, ncol=5, byrow=TRUE)
b

logo <- brick(system.file("external/rlogo.grd", package="raster"))
getValuesBlock(logo, row=35, nrows=3, col=50, ncols=3, lyrs=2:3)

[Package raster version 3.5-2 Index]