match {raster}	R Documentation

Value matching for Raster* objects

Description

match returns a Raster* object with the position of the matched values. The cell values are the index of the table argument.

%in% returns a logical Raster* object indicating if the cells values were matched or not.

Usage


match(x, table, nomatch = NA_integer_, incomparables = NULL)

x %in% table

Arguments

x
Raster* object

table
vector of the values to be matched against

nomatch
the value to be returned in the case when no match is found. Note that it is coerced to integer

incomparables
a vector of values that cannot be matched. Any value in x matching a value in this vector is assigned the nomatch value. For historical reasons, FALSE is equivalent to NULL

Value

Raster* object

See Also

calc, match

Examples


r <- raster(nrow=10, ncol=10)
values(r) <- 1:100
m <- match(r, c(5:10, 50:55))
n <- r %in% c(5:10, 50:55)

[Package raster version 3.5-2 Index]