Distance between points

Calculate the geographic distance between two (sets of) points on the WGS ellipsoid (lonlat=TRUE) or on a plane (lonlat=FALSE). If both sets do not have the same number of points, the distance between each pair of points is given. If both sets have the same number of points, the distance between each point and the corresponding point in the other set is given, except if allpairs=TRUE.

pointDistance(p1, p2, lonlat, allpairs=FALSE, ...)

Arguments

p1

x and y coordinate of first (set of) point(s), either as c(x, y), matrix(ncol=2), or SpatialPoints*.

p2

x and y coordinate of second (set of) second point(s) (like for p1). If this argument is missing, a distance matrix is computed for p1

lonlat

logical. If TRUE, coordinates should be in degrees; else they should represent planar ('Euclidean') space (e.g. units of meters)

allpairs

logical. Only relevant if the number of points in x and y is the same. If FALSE the distance between each point in x with the corresponding point in y is returned. If TRUE a full distance matrix is returned

...

Additional arguments. None implemented

Value

A single value, or a vector, or matrix of values giving the distance in meters (lonlat=TRUE) or map-units (for instance, meters in the case of UTM) If p2 is missing, a distance matrix is returned

See also

Author

Robert J. Hijmans and Jacob van Etten. The distance for longitude/latitude data uses GeographicLib by C.F.F. Karney

Examples

a <- cbind(c(1,5,55,31),c(3,7,20,22))
b <- cbind(c(4,2,8,65),c(50,-90,20,32))   

pointDistance(c(0, 0), c(1, 1), lonlat=FALSE)
#> [1] 1.414214
pointDistance(c(0, 0), c(1, 1), lonlat=TRUE)
#> [1] 156899.6
pointDistance(c(0, 0), a, lonlat=TRUE)
#> [1]  349889.7  952593.2 6383965.5 4151286.9
pointDistance(a, b, lonlat=TRUE)
#> [1]  5216833 10776024  4901277  3534636
   
#Make a distance matrix 
dst <- pointDistance(a, lonlat=TRUE)
# coerce to dist object
dst <- as.dist(dst)