rasterToPolygons {raster}	R Documentation

Raster to polygons conversion

Description

Raster to polygons conversion. Cells with NA are not converted. A function can be used to select a subset of the raster cells (by their values).

Usage


rasterToPolygons(x, fun=NULL, n=4, na.rm=TRUE, digits=12, dissolve=FALSE)

Arguments

x
Raster* object

fun
function to select a subset of raster values (only allowed if x has a single layer)

n
integer. The number of nodes for each polygon. Only 4, 8, and 16 are allowed

na.rm
If TRUE, cells with NA values in all layers are ignored

digits
number of digits to round the coordinates to

dissolve
logical. If TRUE, polygons with the same attribute value will be dissolved into multi-polygon regions. This option requires the rgeos package

Details

fun should be a simple function returning a logical value.

E.g.: fun=function(x){x==1} or fun=function(x){x>3 & x<6}

Value

SpatialPolygonsDataFrame

Examples

 
r <- raster(nrow=18, ncol=36)
values(r) <- runif(ncell(r)) * 10
r[r>8] <- NA
pol <- rasterToPolygons(r, fun=function(x){x>6})

#plot(r > 6)
#plot(pol, add=TRUE, col='red')

[Package raster version 3.5-2 Index]