Extent math {raster}	R Documentation

round Extent coordinates

Description

use round(x, digits=0) to round the coordinates of an Extent object to the number of digits specified. This can be useful when dealing with a small imprecision in the data (e.g. 179.9999 instead of 180). floor and ceiling move the coordiantes to the outer or inner whole integer numbers.

It is also possible to use Arithmetic functions with Extent objects (but these work perhaps unexpectedly!)

See Math-methods for these (and many more) methods with Raster* objects.

Usage


## S4 method for signature 'Extent'
floor(x)
## S4 method for signature 'Extent'
ceiling(x)

Arguments

x
Extent object

See Also

Math-methods

Examples


e <- extent(c(0.999999,  10.000011, -60.4, 60))
round(e)
ceiling(e)
floor(e)

[Package raster version 3.5-2 Index]