sampleRegular {raster}	R Documentation

Regular sample

Description

Take a systematic sample from a Raster* object.

Usage


## S4 method for signature 'Raster'
sampleRegular(x, size, ext=NULL, cells=FALSE, xy=FALSE, asRaster=FALSE, 
            sp=FALSE, useGDAL=FALSE, ...)


Arguments

x
Raster object

size
positive integer giving the number of items to choose.

ext
Extent. To limit regular sampling to the area within that box

cells
logical. Also return sampled cell numbers (if asRaster=FALSE)

xy
logical. If TRUE, coordinates of sampled cells are also returned

asRaster
logical. If TRUE, a RasterLayer or RasterBrick is returned, rather than the sampled values

sp
logical. If TRUE, a SpatialPointsDataFrame is returned

useGDAL
logical. If TRUE, GDAL is used to sample in some cases. This is quicker, but can result in values for a different set of cells than when useGDAL=FALSE. Only for rasters that are accessed via rgdal, and are not rotated. When TRUE arguments cells, xy, and sp are ignored (i.e., FALSE

...
additional arguments. None implemented

Value

A vector (single layer object), matrix (multi-layered object; or if cells=TRUE, or xy=TRUE), Raster* object (if asRaster=TRUE), or SpatialPointsDataFrame (if sp=TRUE)

See Also

sampleRandom, sampleStratified

Examples


 r <- raster(system.file("external/test.grd", package="raster"))
 v <- sampleRegular(r, size=100)
 x <- sampleRegular(r, size=100, asRaster=TRUE)

[Package raster version 3.5-2 Index]