spplot {raster}	R Documentation

Use spplot to plot a Raster* object

Description

A wrapper function around spplot (sp package). With spplot it is easy to map several layers with a single legend for all maps. ssplot is itself a wrapper around the levelplot function in the lattice package, and see the help for these functions for additional options.

One of the advantages of these wrapper functions is the additional maxpixels argument to sample large Raster objects for faster drawing.

Methods

spplot(obj, ..., maxpixels=50000, as.table=TRUE)

obj 	 A Raster* object

... 	 Any argument that can be passed to spplot and levelplot

maxpixels 	 Integer. Number of pixels to sample from each layer of large Raster objects

See Also

plot, plotRGB

The rasterVis package has more advanced plotting methods for Raster objects

Examples

 
r <- raster(system.file("external/test.grd", package="raster"))
s <- stack(r, r*2)
names(s) <- c('meuse', 'meuse x 2')

spplot(s)

pts <- data.frame(sampleRandom(r, 10, xy=TRUE))
coordinates(pts) <- ~ x + y

spplot(s, scales = list(draw = TRUE), 
		xlab = "easting", ylab = "northing", 
		col.regions = rainbow(99, start=.1), 
		names.attr=c('original', 'times two'),
		sp.layout = list("sp.points", pts, pch=20, cex=2, col='black'),
		par.settings = list(fontsize = list(text = 12)), at = seq(0, 4000, 500))

[Package raster version 2.7-15 Index]