stack {raster}	R Documentation

Create a RasterStack object

Description

A RasterStack is a collection of RasterLayer objects with the same spatial extent and resolution. A RasterStack can be created from RasterLayer objects, or from raster files, or both. It can also be created from a SpatialPixelsDataFrame or a SpatialGridDataFrame object.

Usage


## S4 method for signature 'character'
stack(x, ..., bands=NULL, varname="", native=FALSE, RAT=TRUE, quick=FALSE)

## S4 method for signature 'Raster'
stack(x, ..., layers=NULL)

## S4 method for signature 'missing'
stack(x)

## S4 method for signature 'list'
stack(x, bands=NULL, native=FALSE, RAT=TRUE, ...)

Arguments

x
filename (character), Raster* object, missing (to create an empty RasterStack), SpatialGrid*, SpatialPixels*, or list (of filenames and/or Raster* objects). If x is a list, additional arguments ... are ignored

bands
integer. which bands (layers) of the file should be used (default is all layers)

layers
integer (or character with layer names) indicating which layers of a RasterBrick should be used (default is all layers)

native
logical. If TRUE native drivers are used instead of gdal drivers (where available, such as for BIL and Arc-ASCII files)

RAT
logical. If TRUE a raster attribute table is created for files that have one

quick
logical. If TRUE the extent and resolution of the objects are not compared. This speeds up the creation of the RasteStack but should be use with great caution. Only use this option when you are absolutely sure that all the data in all the files are aligned, and you need to create RasterStack for many (>100) files

varname
character. To select the variable of interest in a NetCDF file (see raster)

...
additional filenames or Raster* objects

Value

RasterStack

See Also

addLayer, dropLayer, raster, brick

Examples


# file with one layer
fn <- system.file("external/test.grd", package="raster")
s <- stack(fn, fn)
r <- raster(fn)
s <- stack(r, fn) 
nlayers(s)

# file with three layers
slogo <- stack(system.file("external/rlogo.grd", package="raster")) 
nlayers(slogo)
slogo

[Package raster version 3.5-2 Index]