writeFormats {raster}	R Documentation

File types for writing

Description

List supported file types for writing RasterLayer values to disk.

When a function writes a file to disk, the file format is determined by the 'format=' argument if supplied, or else by the file extension (if the extension is known). If other cases the default format is used. The 'factory-fresh' default format is 'raster', but this can be changed using rasterOptions.

Usage


writeFormats()

Details

writeFormats returns a matrix of the file formats (the "drivers") that are supported.

Supported formats include:

File type 	 Long name 	 default extension 	 Multiband support

raster 	 'Native' raster package format 	 .grd 	 Yes

ascii 	 ESRI Ascii 	 .asc 	 No

SAGA 	 SAGA GIS 	 .sdat 	 No

IDRISI 	 IDRISI 	 .rst 	 No

CDF 	 netCDF (requires ncdf4) 	 .nc 	 Yes

GTiff 	 GeoTiff (requires rgdal) 	 .tif 	 Yes

ENVI 	 ENVI .hdr Labelled 	 .envi 	 Yes

EHdr 	 ESRI .hdr Labelled 	 .bil 	 Yes

HFA 	 Erdas Imagine Images (.img) 	 .img 	 Yes

See Also

GDALDriver-class

Examples

 
writeFormats() 

[Package raster version 3.5-2 Index]