File types for writing

List supported file types for writing RasterLayer values to disk.

When a function writes a file to disk, the file format is determined by the 'format=' argument if supplied, or else by the file extension (if the extension is known). If other cases the default format is used. The 'factory-fresh' default format is 'raster', but this can be changed using rasterOptions.

writeFormats()

Details

writeFormats returns a matrix of the file formats (the "drivers") that are supported.

Supported formats include:

File type	Long name	default extension	Multiband support

raster	'Native' raster package format	.grd	Yes

ascii	ESRI Ascii	.asc	No

SAGA	SAGA GIS	.sdat	No

IDRISI	IDRISI	.rst	No

CDF	netCDF (requires ncdf4)	.nc	Yes

GTiff	GeoTiff	.tif	Yes

ENVI	ENVI .hdr Labelled	.envi	Yes

EHdr	ESRI .hdr Labelled	.bil	Yes

HFA	Erdas Imagine Images (.img)	.img	Yes

Examples

writeFormats() 
#>       name              long_name                                   
#>  [1,] "raster"          "R-raster"                                  
#>  [2,] "SAGA"            "SAGA GIS"                                  
#>  [3,] "IDRISI"          "IDRISI"                                    
#>  [4,] "IDRISIold"       "IDRISI (img/doc)"                          
#>  [5,] "BIL"             "Band by Line"                              
#>  [6,] "BSQ"             "Band Sequential"                           
#>  [7,] "BIP"             "Band by Pixel"                             
#>  [8,] "ascii"           "Arc ASCII"                                 
#>  [9,] "CDF"             "NetCDF"                                    
#> [10,] "AAIGrid"         "Arc/Info ASCII Grid"                       
#> [11,] "ADRG"            "ARC Digitized Raster Graphics"             
#> [12,] "ARG"             "Azavea Raster Grid format"                 
#> [13,] "BAG"             "Bathymetry Attributed Grid"                
#> [14,] "BLX"             "Magellan topo (.blx)"                      
#> [15,] "BMP"             "MS Windows Device Independent Bitmap"      
#> [16,] "BT"              "VTP .bt (Binary Terrain) 1.3 Format"       
#> [17,] "BYN"             "Natural Resources Canada's Geoid"          
#> [18,] "CALS"            "CALS (Type 1)"                             
#> [19,] "COG"             "Cloud optimized GeoTIFF generator"         
#> [20,] "CTable2"         "CTable2 Datum Grid Shift"                  
#> [21,] "DTED"            "DTED Elevation Raster"                     
#> [22,] "EHdr"            "ESRI .hdr Labelled"                        
#> [23,] "ELAS"            "ELAS"                                      
#> [24,] "ENVI"            "ENVI .hdr Labelled"                        
#> [25,] "ERS"             "ERMapper .ers Labelled"                    
#> [26,] "FIT"             "FIT Image"                                 
#> [27,] "GIF"             "Graphics Interchange Format (.gif)"        
#> [28,] "GPKG"            "GeoPackage"                                
#> [29,] "GRIB"            "GRIdded Binary (.grb, .grb2)"              
#> [30,] "GS7BG"           "Golden Software 7 Binary Grid (.grd)"      
#> [31,] "GSAG"            "Golden Software ASCII Grid (.grd)"         
#> [32,] "GSBG"            "Golden Software Binary Grid (.grd)"        
#> [33,] "GTX"             "NOAA Vertical Datum .GTX"                  
#> [34,] "GTiff"           "GeoTIFF"                                   
#> [35,] "HDF4Image"       "HDF4 Dataset"                              
#> [36,] "HF2"             "HF2/HFZ heightfield raster"                
#> [37,] "HFA"             "Erdas Imagine Images (.img)"               
#> [38,] "IDA"             "Image Data and Analysis"                   
#> [39,] "ILWIS"           "ILWIS Raster Map"                          
#> [40,] "INGR"            "Intergraph Raster"                         
#> [41,] "ISCE"            "ISCE raster"                               
#> [42,] "ISIS2"           "USGS Astrogeology ISIS cube (Version 2)"   
#> [43,] "ISIS3"           "USGS Astrogeology ISIS cube (Version 3)"   
#> [44,] "JP2OpenJPEG"     "JPEG-2000 driver based on OpenJPEG library"
#> [45,] "JPEG"            "JPEG JFIF"                                 
#> [46,] "KMLSUPEROVERLAY" "Kml Super Overlay"                         
#> [47,] "KRO"             "KOLOR Raw"                                 
#> [48,] "LAN"             "Erdas .LAN/.GIS"                           
#> [49,] "LCP"             "FARSITE v.4 Landscape File (.lcp)"         
#> [50,] "Leveller"        "Leveller heightfield"                      
#> [51,] "MBTiles"         "MBTiles"                                   
#> [52,] "MRF"             "Meta Raster Format"                        
#> [53,] "NGW"             "NextGIS Web"                               
#> [54,] "NITF"            "National Imagery Transmission Format"      
#> [55,] "NTv2"            "NTv2 Datum Grid Shift"                     
#> [56,] "NWT_GRD"         "Northwood Numeric Grid Format .grd/.tab"   
#> [57,] "PAux"            "PCI .aux Labelled"                         
#> [58,] "PCIDSK"          "PCIDSK Database File"                      
#> [59,] "PCRaster"        "PCRaster Raster File"                      
#> [60,] "PDF"             "Geospatial PDF"                            
#> [61,] "PDS4"            "NASA Planetary Data System 4"              
#> [62,] "PNG"             "Portable Network Graphics"                 
#> [63,] "PNM"             "Portable Pixmap Format (netpbm)"           
#> [64,] "PostGISRaster"   "PostGIS Raster driver"                     
#> [65,] "R"               "R Object Data Store"                       
#> [66,] "RMF"             "Raster Matrix Format"                      
#> [67,] "ROI_PAC"         "ROI_PAC raster"                            
#> [68,] "RRASTER"         "R Raster"                                  
#> [69,] "RST"             "Idrisi Raster A.1"                         
#> [70,] "Rasterlite"      "Rasterlite"                                
#> [71,] "SAGA"            "SAGA GIS Binary Grid (.sdat, .sg-grd-z)"   
#> [72,] "SGI"             "SGI Image File Format 1.0"                 
#> [73,] "SIGDEM"          "Scaled Integer Gridded DEM .sigdem"        
#> [74,] "SRTMHGT"         "SRTMHGT File Format"                       
#> [75,] "Terragen"        "Terragen heightfield"                      
#> [76,] "USGSDEM"         "USGS Optional ASCII DEM (and CDED)"        
#> [77,] "VICAR"           "MIPL VICAR file"                           
#> [78,] "WEBP"            "WEBP"                                      
#> [79,] "WMS"             "OGC Web Map Service"                       
#> [80,] "WMTS"            "OGC Web Map Tile Service"                  
#> [81,] "XPM"             "X11 PixMap Format"                         
#> [82,] "XYZ"             "ASCII Gridded XYZ"                         
#> [83,] "ZMap"            "ZMap Plus Grid"                            
#> [84,] "Zarr"            "Zarr"