Write raster data to a file

Write an entire Raster* object to a file, using one of the many supported formats. See writeValues for writing in chunks (e.g. by row).

When writing a file to disk, the file format is determined by the 'format=' argument if supplied, or else by the file extension (if the extension is known). If other cases the default format is used. The default format is 'raster', but this setting can be changed (see rasterOptions).

# S4 method for RasterLayer,character
writeRaster(x, filename, format, ...)

# S4 method for RasterStackBrick,character
writeRaster(x, filename, format, bylayer, suffix='numbers', ...)

Arguments

x

Raster* object

filename

Output filename

format

Character. Output file type. See writeFormats. If this argument is not provided, it is attempted to infer it from the filename extension. If that fails, the default format is used. The default format is 'raster', but this can be changed using rasterOptions

...

Additional arguments:

datatypeCharacter. Output data type (e.g. 'INT2S' or 'FLT4S'). See dataType. If no datatype is specified, 'FLT4S' is used, unless this default value was changed with rasterOptions

overwrite: Logical. If TRUE, "filename" will be overwritten if it exists

progress: Character. Set a value to show a progress bar. Valid values are "text" and "window".

NAflag: Numeric. To overwrite the default value used to represent NA in a file

bandorder: Character. 'BIL', 'BIP', or 'BSQ'. For 'native' file formats only. For some other formats you can use the 'options' argument (see below)

options: Character. File format specific GDAL options. E.g., when writing a geotiff file you can use: options=c("COMPRESS=NONE", "TFW=YES")

You can use options=c("PROFILE=BASELINE") to create a plain tif with no GeoTIFF tags. This can be useful when writing files to be read by applications intolerant of unrecognised tags.

NetCDF files have the following additional, optional, arguments: varname, varunit, longname, xname, yname, zname, zunit

prj: Logical. If TRUE, the crs is written to a .prj file. This can be useful when writing to an ascii file or another file type that does not store the crs

setStatistics: logical. If TRUE (the default) the min and max cell values are written to file (if the format permits it)

bylayer

if TRUE, write a separate file for each layer. You can provide a vector of filenames that matches the number of layers. Or you can provide a single filename that will get a unique suffix (see below)

suffix

'numbers' or 'names' to determine the suffix that each file gets when bylayer=TRUE; either a number between 1 and nlayers(x) or names(x)

Details

See writeFormats for supported file types ("formats", "drivers").

In multi-layer files (i.e. files saved from RasterStack or RasterBrick objects), in the native 'raster' format, the band-order can be set to BIL ('Bands Interleaved by Line'), BIP ('Bands Interleaved by Pixels') or BSQ ('Bands SeQuential'). Note that bandorder is not the same as filetype here.

Supported file types include:

File type	Long name	default extension	Multiband support

raster	'Native' raster package format	.grd	Yes

ascii	ESRI Ascii	.asc	No

SAGA	SAGA GIS	.sdat	No

IDRISI	IDRISI	.rst	No

CDF	netCDF (requires ncdf4)	.nc	Yes

GTiff	GeoTiff	.tif	Yes

ENVI	ENVI .hdr Labelled	.envi	Yes

EHdr	ESRI .hdr Labelled	.bil	Yes

HFA	Erdas Imagine Images (.img)	.img	Yes

Value

This function is used for the side-effect of writing values to a file.

See also

Examples

tmp <- tempdir()
r <- raster(system.file("external/test.grd", package="raster"))

# take a small part
r <- crop(r, extent(179880, 180800, 329880, 330840) )

# write to an integer binary file
rf <- writeRaster(r, filename=file.path(tmp, "allint.grd"), datatype='INT4S', overwrite=TRUE)

# make a brick and save multi-layer file
b <- brick(r, sqrt(r))
bf <- writeRaster(b, filename=file.path(tmp, "multi.grd"), bandorder='BIL', overwrite=TRUE)

# write to a new geotiff file
rf <- writeRaster(r, filename=file.path(tmp, "test.tif"), format="GTiff", overwrite=TRUE)
bf <- writeRaster(b, filename=file.path(tmp, "multi.tif"), 
            options="INTERLEAVE=BAND", overwrite=TRUE)

 
# write to netcdf 
if (require(ncdf4)) {  
 rnc <- writeRaster(r, filename=file.path(tmp, "netCDF.nc"), format="CDF", overwrite=TRUE)   
}
#> Loading required package: ncdf4